fix(client): improved mouse navigation while searching, and fixed an infinite loop scenario
I solved a bug that used to occur while performing a search and receiving new data from the server
at the same time. This caused the UI to refresh, and reapply the search. But reapplying the search
caused the UI to request new information from the server. Hence triggering a new cycle of reapplying
search, etc. I  I got rid of that infinite loop by not reapplying search when new data was received
from the network.