feat(project): added support for an Overview panel showing the server, hosts, and agents all at once
There is now an Overview panel (with a default opening on launch) enabling you to see all your
resources at a glance, whether they are on the main server, on a remote host, or on a connected
agent. You can also pick any of them using your keyboard or your mouse. Ultimately, this feature is
100% responsive for tablet and mobile.